Move `getUpdateCommand` to util
