+++ src/commands/update.ts
@@ -1,15 +1,11 @@
 import chalk from 'chalk';
-import { Stats } from 'fs';
-import { dirname, join, resolve } from 'path';
-import { readJSON, lstat, readlink } from 'fs-extra';
 
-import cmd from '../util/output/cmd';
 import logo from '../util/output/logo';
 import handleError from '../util/handle-error';
 import getArgs from '../util/get-args';
 import { NowContext } from '../types';
 import createOutput from '../util/output';
-import { version } from '../../package.json';
+import getUpdateCommand from '../util/get-update-command';
 
 const help = () => {
   console.log(`
@@ -35,31 +31,6 @@ const help = () => {
   `);
 };
 
-// `npm` tacks a bunch of extra properties on the `package.json` file,
-// so check for one of them to determine yarn vs. npm.
-async function isYarn(): Promise<boolean> {
-  let s: Stats;
-  let binPath = process.argv[1];
-  while (true) {
-    s = await lstat(binPath);
-    if (s.isSymbolicLink()) {
-      binPath = resolve(dirname(binPath), await readlink(binPath));
-    } else {
-      break;
-    }
-  }
-  const pkgPath = join(dirname(binPath), '..', 'package.json');
-  const pkg = await readJSON(pkgPath);
-  return !('_id' in pkg);
-}
-
-export async function getUpgradeCommand(): Promise<string> {
-  const tag = version.includes('canary') ? 'canary' : 'latest';
-  return (await isYarn())
-    ? `Please run ${cmd(`yarn global add now@${tag}`)} to update Now CLI.`
-    : `Please run ${cmd(`npm install -g now@${tag}`)} to update Now CLI.`;
-}
-
 export default async function main(ctx: NowContext): Promise<number> {
   let argv;
 
@@ -84,6 +55,6 @@ export default async function main(ctx: NowContext): Promise<number> {
 
   const debugEnabled = argv['--debug'];
   const output = createOutput({ debug: debugEnabled });
-  output.log(await getUpgradeCommand());
+  output.log(await getUpdateCommand());
   return 0;
 }
